payment success

+++ payment/models.py
@@ -4,7 +4,7 @@ from order.models import Order
 # Create your models here.
 class Payment(models.Model):
     order = models.ForeignKey(Order,related_name = "payment",on_delete = models.PROTECT)
-    ammount = models.PositiveIntegerField()
+    ammount = models.FloatField()
     payment_mode = models.CharField(max_length = 100,choices = (('khalti','Khalti'),('esewa','Esewa'),('cod','Cash On Ddelivery')))
     refrence_id = models.CharField(max_length = 4000,default = 'cod')
     remarks = models.CharField(max_length = 3000,null = True,blank = True)
